Enhance server scrolling demo (740#issuecomment-310806538)
We enhance "infinite" server-side scrolling example
by binding to (scroll) instead of (page) events.

We can now detect when the user scrolls down to the bottom
of the list, and load only as many results as needed,
until the user scrolls all the way down again.

We also add some styles to dynamically resize the list
according to user's viewport size, and to display
the progress indicator at the bottom of the list.

Overall, this provides a more natural "infinite" scrolling experience.